requirements/base.txt
django
djangorestframework
django-tagging
#django_filter
django-storages
numpy
pymongo
ansible
#https://github.com/neo4j-contrib/django-neomodel
django_neomodel
psycopg2-binary
networkx
pika
#django-celery-results
django-filter

